Explicitly bind HTTP API server to 127.0.0.1
Also make it use the --port (env.Port) command line option. The motivation
for limiting API server to 127.0.0.1 is one of the ways to make sure
the server is running locally. For many reasons, you don't really want
to run it elsewhere as it will be holding your private keys.